create dfcandle model, and move cwcandle struct from usecase to models
